Visualizing The Gender Gap In College Degrees
In this project, I'll extend on previous work: visualizing the gender gap across college degrees.

There was a focus on STEM degrees but now I will generate line charts to compare across all degree categories. In the last step of this project, I'll explore how to export the final diagram I create as an image file.